backend/requirements.txt
# Core FastAPI backend
fastapi
uvicorn[standard]
pydantic
tinydb
python-multipart
python-jose
passlib[bcrypt]
dramatiq[redis]
redis
python-dotenv
fhir.resources

# Observability / Logging
structlog
sentry-sdk
prometheus-fastapi-instrumentator

# OpenTelemetry for tracing
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
